test_requirements.txt
flake8==2.6.0
mock==2.0.0
mongomock==3.4.0
testing.postgresql==1.3.0
